+++ realm/realm-library/src/main/java/io/realm/internal/ObjectServerFacade.java
@@ -88,4 +88,8 @@ public class ObjectServerFacade {
     // If no session yet exists for this path. Wrap a new Java Session around an existing OS one.
     public void wrapObjectStoreSessionIfRequired(RealmConfiguration config) {
     }
+
+    public void downloadServerChangesIfNeeded(RealmConfiguration config) {
+        // Do nothing
+    }
 }